shuffling code around to fix circular import situation
