+++ tests/test_full_suite_nb.py
@@ -18,7 +18,7 @@ class TestFullTestSuiteNB(unittest.TestCase):
         log_metadata=unittest.mock.DEFAULT,
         log_metrics=unittest.mock.DEFAULT,
         log_test_result=unittest.mock.DEFAULT,
-        _log_input=unittest.mock.DEFAULT,
+        log_input=unittest.mock.DEFAULT,
     )
     @patch("validmind.client.log_input", return_value="1234")
     def test_run_full_suite(self, mock_ipython, *mocks, **mocks2):
