+++ validmind/ai/test_descriptions.py
@@ -2,16 +2,20 @@
 # See the LICENSE file in the root of this repository for details.
 # SPDX-License-Identifier: AGPL-3.0 AND ValidMind Commercial
 
-import concurrent.futures
 import os
+from concurrent.futures import ThreadPoolExecutor
 
-from openai import AzureOpenAI, Client, OpenAI
+from validmind.utils import md_to_html
 
-from .api_client import get_ai_key
-from .logging import get_logger
+from ..logging import get_logger
+
+__executor = ThreadPoolExecutor()
 
 logger = get_logger(__name__)
 
+
+AI_REVISION_NAME = "Generated by ValidMind AI"
+DEFAULT_REVISION_NAME = "Default Description"
 
 SYSTEM_PROMPT = """
 You are an expert data scientist and MRM specialist.
@@ -79,67 +83,6 @@ Test ID: `{test_name}`
 The attached plots show the results of the test.
 """.strip()
 
-__client = None
-__model = None
-
-# can be None, True or False (ternary to represent initial state, ack and failed ack)
-__ack = None
-
-__executor = concurrent.futures.ThreadPoolExecutor()
-
-
-def get_client_and_model():
-    """Get model and client to use for generating interpretations
-
-    On first call, it will look in the environment for the API key endpoint, model etc.
-    and store them in a global variable to avoid loading them up again.
-    """
-    global __client, __model
-
-    if __client and __model:
-        return __client, __model
-
-    if "OPENAI_API_KEY" in os.environ:
-        __client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))
-        __model = os.getenv("VM_OPENAI_MODEL", "gpt-4o")
-
-        logger.debug(f"Using OpenAI {__model} for generating descriptions")
-
-    elif "AZURE_OPENAI_KEY" in os.environ:
-        if "AZURE_OPENAI_ENDPOINT" not in os.environ:
-            raise ValueError(
-                "AZURE_OPENAI_ENDPOINT must be set to run LLM tests with Azure"
-            )
-
-        if "AZURE_OPENAI_MODEL" not in os.environ:
-            raise ValueError(
-                "AZURE_OPENAI_MODEL must be set to run LLM tests with Azure"
-            )
-
-        __client = AzureOpenAI(
-            azure_endpoint=os.getenv("AZURE_OPENAI_ENDPOINT"),
-            api_key=os.getenv("AZURE_OPENAI_KEY"),
-            api_version=os.getenv("AZURE_OPENAI_VERSION", "2023-05-15"),
-        )
-        __model = os.getenv("AZURE_OPENAI_MODEL")
-
-        logger.debug(f"Using Azure OpenAI {__model} for generating descriptions")
-
-    else:
-        try:
-            response = get_ai_key()
-            __client = Client(base_url=response["url"], api_key=response["key"])
-            __model = "gpt-4o"  # TODO: backend should tell us which model to use
-            logger.debug(f"Using ValidMind {__model} for generating descriptions")
-        except Exception as e:
-            logger.debug(f"Failed to get API key: {e}")
-            raise ValueError(
-                "OPENAI_API_KEY, AZURE_OPENAI_KEY must be set, or your account "
-                "must be setup to use ValidMind's LLM in order to use LLM features"
-            )
-
-    return __client, __model
-
 
 class DescriptionFuture:
     """This will be immediately returned from generate_description so that
@@ -154,8 +97,6 @@ class DescriptionFuture:
         self._future = future
 
     def get_description(self):
-        from .utils import md_to_html
-
         if isinstance(self._future, str):
             description = self._future
         else:
@@ -175,7 +116,11 @@ def generate_description(
     if not test_summary and not figures:
         raise ValueError("No summary or figures provided - cannot generate description")
 
+    # TODO: fix circular import
+    from validmind.ai.utils import get_client_and_model
+
     client, model = get_client_and_model()
+
     # get last part of test id
     test_name = test_id.split(".")[-1]
     # truncate the test description to save time
@@ -267,26 +212,61 @@ def background_generate_description(
     return DescriptionFuture(__executor.submit(wrapped))
 
 
-def is_configured():
-    global __ack
+def get_description_metadata(
+    test_id,
+    default_description,
+    summary=None,
+    figures=None,
+    prefix="metric_description",
+):
+    """Get Metadata Dictionary for a Test or Metric Result
 
-    if __ack:
-        return True
+    Generates an LLM interpretation of the test results or uses the default
+    description and returns a metadata object that can be logged with the test results.
 
-    try:
-        client, model = get_client_and_model()
-        # send an empty message with max_tokens=1 to "ping" the API
-        response = client.chat.completions.create(
-            model=model,
-            messages=[{"role": "user", "content": ""}],
-            max_tokens=1,
+    By default, the description is generated by an LLM that will interpret the test
+    results and provide a human-readable description. If the summary or figures are
+    not provided, or the `VALIDMIND_LLM_DESCRIPTIONS_ENABLED` environment variable is
+    set to `0` or `false` or no LLM has been configured, the default description will
+    be used as the test result description.
+
+    Note: Either the summary or figures must be provided to generate the description.
+
+    Args:
+        test_id (str): The test ID
+        default_description (str): The default description for the test
+        summary (Any): The test summary or results to interpret
+        figures (List[Figure]): The figures to attach to the test suite result
+        prefix (str): The prefix to use for the content ID (Default: "metric_description")
+
+    Returns:
+        dict: The metadata object to be logged with the test results
+    """
+    env_disabled = os.getenv("VALIDMIND_LLM_DESCRIPTIONS_ENABLED", "1") in [
+        "0",
+        "false",
+    ]
+
+    # TODO: fix circular import
+    from validmind.ai.utils import is_configured
+
+    if (summary or figures) and not env_disabled and is_configured():
+        revision_name = AI_REVISION_NAME
+
+        # get description future and set it as the description in the metadata
+        # this will lazily retrieved so it can run in the background in parallel
+        description = background_generate_description(
+            test_id=test_id,
+            test_description=default_description,
+            test_summary=summary,
+            figures=figures,
         )
-        logger.debug(
-            f"Received response from OpenAI: {response.choices[0].message.content}"
-        )
-        __ack = True
-    except Exception as e:
-        logger.debug(f"Failed to connect to OpenAI: {e}")
-        __ack = False
 
-    return __ack
+    else:
+        revision_name = DEFAULT_REVISION_NAME
+        description = md_to_html(default_description, mathml=True)
+
+    return {
+        "content_id": f"{prefix}:{test_id}::{revision_name}",
+        "text": description,
+    }
