+++ validmind/unit_metrics/composite.py
@@ -6,9 +6,10 @@ from dataclasses import dataclass
 from typing import List, Tuple, Union
 from uuid import uuid4
 
+from ..ai.test_descriptions import get_description_metadata
 from ..logging import get_logger
 from ..tests.decorator import _inspect_signature
-from ..utils import get_description_metadata, run_async, test_id_to_name
+from ..utils import run_async, test_id_to_name
 from ..vm_models.test.metric import Metric
 from ..vm_models.test.metric_result import MetricResult
 from ..vm_models.test.result_summary import ResultSummary, ResultTable
