+++ validmind/utils.py
@@ -6,7 +6,6 @@ import asyncio
 import difflib
 import json
 import math
-import os
 import re
 import sys
 from platform import python_version
@@ -26,11 +25,8 @@ from matplotlib.axes._axes import _log as matplotlib_axes_logger
 from numpy import ndarray
 from tabulate import tabulate
 
-from .ai import background_generate_description, is_configured
 from .html_templates.content_blocks import math_jax_snippet, python_syntax_highlighting
-
-AI_REVISION_NAME = "Generated by ValidMind AI"
-DEFAULT_REVISION_NAME = "Default Description"
+from .logging import get_logger
 
 DEFAULT_BIG_NUMBER_DECIMALS = 2
 DEFAULT_SMALL_NUMBER_DECIMALS = 4
@@ -52,6 +48,8 @@ params = {
 }
 pylab.rcParams.update(params)
 #################################
+
+logger = get_logger(__name__)
 
 
 def is_notebook() -> bool:
@@ -460,60 +458,3 @@ def md_to_html(md: str, mathml=False) -> str:
     )
 
     return html
-
-
-def get_description_metadata(
-    test_id,
-    default_description,
-    summary=None,
-    figures=None,
-    prefix="metric_description",
-):
-    """Get Metadata Dictionary for a Test or Metric Result
-
-    Generates an LLM interpretation of the test results or uses the default
-    description and returns a metadata object that can be logged with the test results.
-
-    By default, the description is generated by an LLM that will interpret the test
-    results and provide a human-readable description. If the summary or figures are
-    not provided, or the `VALIDMIND_LLM_DESCRIPTIONS_ENABLED` environment variable is
-    set to `0` or `false` or no LLM has been configured, the default description will
-    be used as the test result description.
-
-    Note: Either the summary or figures must be provided to generate the description.
-
-    Args:
-        test_id (str): The test ID
-        default_description (str): The default description for the test
-        summary (Any): The test summary or results to interpret
-        figures (List[Figure]): The figures to attach to the test suite result
-        prefix (str): The prefix to use for the content ID (Default: "metric_description")
-
-    Returns:
-        dict: The metadata object to be logged with the test results
-    """
-    env_disabled = os.getenv("VALIDMIND_LLM_DESCRIPTIONS_ENABLED", "1") in [
-        "0",
-        "false",
-    ]
-
-    if (summary or figures) and not env_disabled and is_configured():
-        revision_name = AI_REVISION_NAME
-
-        # get description future and set it as the description in the metadata
-        # this will lazily retrieved so it can run in the background in parallel
-        description = background_generate_description(
-            test_id=test_id,
-            test_description=default_description,
-            test_summary=summary,
-            figures=figures,
-        )
-
-    else:
-        revision_name = DEFAULT_REVISION_NAME
-        description = md_to_html(default_description, mathml=True)
-
-    return {
-        "content_id": f"{prefix}:{test_id}::{revision_name}",
-        "text": description,
-    }
